requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.18
aiosignal==1.3.2
alembic==1.13.1
annotated-types==0.7.0
anyio==4.9.0
appnope==0.1.4
asttokens==3.0.0
attrs==25.3.0
certifi==2025.1.31
click==8.1.8
comm==0.2.2
cramjam==2.10.0
debugpy==1.8.14
decorator==5.2.1
executing==2.2.0
fastapi==0.109.2
fastparquet==2024.11.0
frozenlist==1.6.0
fsspec==2025.3.2
h11==0.14.0
httpcore==1.0.8
httpx==0.26.0
idna==3.10
ipykernel==6.29.5
ipython==9.1.0
ipython_pygments_lexers==1.1.1
jedi==0.19.2
jupyter_client==8.6.3
jupyter_core==5.7.2
Mako==1.3.10
MarkupSafe==3.0.2
matplotlib-inline==0.1.7
multidict==6.4.3
nest-asyncio==1.6.0
numpy==2.2.5
packaging==25.0
pandas==2.2.3
parso==0.8.4
pexpect==4.9.0
platformdirs==4.3.7
prompt_toolkit==3.0.51
propcache==0.3.1
psutil==7.0.0
psycopg2-binary==2.9.9
ptyprocess==0.7.0
pure_eval==0.2.3
pyarrow==19.0.1
pydantic==2.6.1
pydantic_core==2.16.2
Pygments==2.19.1
python-dateutil==2.8.2
python-dotenv==1.0.1
pytz==2025.2
pyzmq==26.4.0
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.27
# Editable install with no version control (stablecoins==0.1.0)
-e /Users/nvtrang/gblabs/stablecoins
stack-data==0.6.3
starlette==0.36.3
tornado==6.4.2
traitlets==5.14.3
typing_extensions==4.13.2
tzdata==2025.2
uvicorn==0.27.1
wcwidth==0.2.13
yarl==1.20.0
